Match more flags to get runnable packages
AppHibernationService was failing to query a few edge cases which
caused a mismatch between the PackageManager and
AppHibernationManager version of all installed packages on the
device.

Specifically, this included apps that are hidden until officially
"installed" (from what I can tell, this seems to be used for system
carrier apps that update to the play version of the APK later) and
packages that are disabled until used.

Bug: 188711013
Test: builds
Change-Id: I5e7a7557d4ee4ebcb848890d97e4b674dd6e18a7