Only load the overflow data once
We would also try loading the overflow data if it was empty
but this works poorly if it's becoming empty due to an app
cancel because we could load contents that are about to be
removed.

I think this can safely work if we just load it based on
the flag.

Adds a test related to this that was only in BubblesTest to
 NewNotifPipelineBubblesTest as well

Test: atest BubblesTest NewNotifPipelineBubblesTest
Test: - have nothing in bubble overflow
      - post 3 bubbles, cancel all, get a bubble, look in the overflow
      => the 3 bubbles shouldn't be there
Bug: 193560795
Change-Id: I1ec2d637281807ae3a2ec7430cff6627b074fa78