Allow to transfer starting window from dying activity
If a launching activity with starting window is died and there
is an activity in the same task, the starting window may be
removed before the next activity is drawn. The symptom will be
obvious if the next activity is slow to draw (black screen).

Another symptom is that if the launching activity requested
different orientation from the display, its rotation transform
will be canceled when removing the activity. That causes the
starting window shows in previous orientation temporally.

Bug: 191429072
Test: Start a new activity on a standalone process into an
      existing task. The activity calls finish and kills
      itself in onCreate.
Change-Id: Icc687854f9885389aa436b617c4bcb231a74c42c