Annotate TextServicesManager with @UserHandleAware
This is a follow up CL to my CL [1], enabled TextServicesManager to
determine the target user ID by Context#getUser().

This CL aims to clarify what methods in TextServicesManager follow
this semantics.

Note that @UserHandleAware is still a kind of hint for readers.  There
should be no change at the runtime.

 [1]: I31dda3ae8795190d44b0622b8335c34ddbc5dd48
      fae54c7e6eb06d4f6330cd2715eb5b187a5c8a37

Bug: 192412909
Test: presubmit
Change-Id: I00bd99ef9d83a13bd60a5aefd22094725b67e271